third_party/blink/web_tests/wpt_internal/private-aggregation/test-internals.window-expected.txt

This is a testharness.js-based test.
[FAIL] CborParser rejects input with a trailing, unparsed byte
  assert_equals: CborParser did not consume trailing bytes:\n42\n expected 0 but got 1
Harness: the test ran to completion.
